+++ packages/testkits/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ts-linq/testkits
 
+## 7.0.33
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.11.0
+  - @ts-linq/provider-postgres@3.0.33
+  - @ts-linq/provider-mysql@3.0.33
+  - @ts-linq/provider-mssql@3.0.33
+  - @ts-linq/core@3.5.2
+  - @ts-linq/metadata@4.1.10
+
 ## 7.0.32
 
 ### Patch Changes
@@ -11,6 +23,7 @@
   - @ts-linq/provider-mssql@3.0.32
   - @ts-linq/provider-mysql@3.0.32
   - @ts-linq/provider-postgres@3.0.32
+
 ## 7.0.31
 
 ### Patch Changes